Move SearchTypeValues and SearchType to global files
