--- src/containers/SearchPage/interfaces.ts
@@ -1,2 +1,0 @@
-export const SearchTypeValues = ['content', 'audio', 'image', 'concept', 'podcast-series'] as const;
-export type SearchType = typeof SearchTypeValues[number];
